+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0 (Next)
 
+* Added a command-line `hyperclient-mcp` tool - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2025/09/12)